map is created from image
